# api/openapi/access-service.yaml
openapi: 3.0.3
info:
  title: OpenIDX Access Service
  description: |
    Zero Trust access proxy with route management, session control, OpenZiti network overlay, posture checks, BrowZer, Guacamole integration, and Privileged Access Management (vault-backed entries, brokered launch with credential injection, checkout controls, session moderation, SSH CA, cloud JIT, Windows app delivery).

    Every route cmd/access-service mounts appears here, and nothing that is not
    mounted does: test/openapi/spec_census_test.go compares this file against the
    live route table in both directions. Operations carry an `x-openidx-handler`
    extension naming the Go handler that serves them; where a payload shape is not
    yet pinned the body is shown as a free-form JSON object rather than invented.
  version: 1.34.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8007
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/access/routes:
    get:
      tags: [Routes]
      summary: List proxy routes
      operationId: listRoutes
      parameters:
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated list of proxy routes
          headers:
            X-Total-Count:
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProxyRoute'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Routes]
      summary: Create proxy route
      operationId: createRoute
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProxyRouteCreate'
      responses:
        '201':
          description: Route created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProxyRoute'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/access/routes/{id}:
    get:
      tags: [Routes]
      summary: Get proxy route
      operationId: getRoute
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Route details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProxyRoute'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Routes]
      summary: Update proxy route
      operationId: updateRoute
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProxyRouteCreate'
      responses:
        '200':
          description: Route updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProxyRoute'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Routes]
      summary: Delete proxy route
      operationId: deleteRoute
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Route deleted
        '404':
          $ref: '#/components/responses/NotFound'

  # --- Sessions ---
...
  /api/v1/access/ziti/fabric/routers/{id}:
    get:
      tags: [Ziti Fabric]
      summary: Get edge router
      operationId: getEdgeRouter
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Edge router details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EdgeRouter'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object

    # --- Auth ---
...
    ProxyRoute:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        from_url:
          type: string
        to_url:
          type: string
        preserve_host:
          type: boolean
        require_auth:
          type: boolean
        allowed_roles:
          type: array
          items:
            type: string
        allowed_groups:
          type: array
          items:
            type: string
        policy_ids:
          type: array
          items:
            type: string
        idle_timeout:
          type: integer
          description: Idle timeout in seconds
        absolute_timeout:
          type: integer
          description: Absolute timeout in seconds
        cors_allowed_origins:
          type: array
          items:
            type: string
        custom_headers:
          type: object
          additionalProperties:
            type: string
        enabled:
          type: boolean
        priority:
          type: integer
        ziti_enabled:
          type: boolean
        ziti_service_name:
          type: string
        idp_id:
          type: string
        route_type:
          type: string
        remote_host:
          type: string
        remote_port:
          type: integer
        reverify_interval:
          type: integer
          description: Re-verification interval in seconds
        posture_check_ids:
          type: array
          items:
            type: string
        inline_policy:
          type: object
        require_device_trust:
          type: boolean
        allowed_countries:
          type: array
          items:
            type: string
        max_risk_score:
          type: number
        guacamole_connection_id:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProxyRouteCreate:
      type: object
      required: [name, from_url, to_url]
      properties:
        name:
          type: string
        description:
          type: string
        from_url:
          type: string
        to_url:
          type: string
        preserve_host:
          type: boolean
        require_auth:
          type: boolean
        allowed_roles:
          type: array
          items:
            type: string
        allowed_groups:
          type: array
          items:
            type: string
        policy_ids:
          type: array
          items:
            type: string
        idle_timeout:
          type: integer
        absolute_timeout:
          type: integer
        cors_allowed_origins:
          type: array
          items:
            type: string
        custom_headers:
          type: object
          additionalProperties:
            type: string
        enabled:
          type: boolean
          default: true
        priority:
          type: integer
        ziti_enabled:
          type: boolean
        ziti_service_name:
          type: string
        idp_id:
          type: string
        route_type:
          type: string
        remote_host:
          type: string
        remote_port:
          type: integer
        reverify_interval:
          type: integer
        posture_check_ids:
          type: array
          items:
            type: string
        inline_policy:
          type: object
        require_device_trust:
          type: boolean
        allowed_countries:
          type: array
          items:
            type: string
        max_risk_score:
          type: number
        guacamole_connection_id:
          type: string

    # --- Sessions ---
...
    EdgeRouter:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        hostname:
          type: string
        is_online:
          type: boolean
        is_verified:
          type: boolean
        created_at:
          type: string
          format: date-time
...